* Changed the ctor for rt_triple to a non-template version.
* Made torrent::Rate::rate_type uint64_t and added sanity checks due to reports of bad download/upload values being reported to some trackers.


git-svn-id: svn://rakshasa.no/libtorrent/trunk/libtorrent@1103 e378c898-3ddf-0310-93e7-cc216c733640

+++ src/torrent/rate.h
@@ -49,7 +49,7 @@ namespace torrent {
 class LIBTORRENT_EXPORT Rate {
 public:
   typedef int32_t                          timer_type;
-  typedef uint32_t                         rate_type;
+  typedef uint64_t                         rate_type;
   typedef uint64_t                         total_type;
 
   typedef std::pair<timer_type, rate_type> value_type;
